requirements.txt
streamlit
pandas
plotly
newsapi-python
transformers
python-dotenv
torch
torchvision
torchaudio